refactor(GameObject): move out from PVZ.h
